chore: Reorganized source and test directories.
