+++ tests/test_data/4xxx.kicad_sym
