+++ tests/test_data/Amplifier_Operational.lib
