+++ tests/test_data/Device.kicad_sym
