+++ tests/test_data/MCU_Microchip_PIC10.lib
