+++ tests/test_data/SpiceLib/sym/NCP1117.slb
