+++ tests/test_data/power.lib
