+++ tox.ini
@@ -34,7 +34,7 @@ allowlist_externals =
     sphinx-build
 commands =
     pip install -e {toxinidir}
-    py.test -m "not spice" --tb=short --color=yes
+    py.test --ignore {toxinidir}/tests/examples -m "not spice" --tb=short --color=yes
 
 [testenv:spice]
 basepython = python3.10
@@ -55,7 +55,7 @@ allowlist_externals =
     sphinx-build
 commands =
     pip install -e {toxinidir}
-    py.test -m spice --tb=short --color=yes
+    py.test --ignore {toxinidir}/tests/examples -m spice --tb=short --color=yes
 
 [coverage:run]
 branch = True
@@ -79,7 +79,7 @@ deps =
 changedir = {toxinidir}/tests
 commands =
     pip install -e {toxinidir}
-    py.test -m "spice" --cov-config={toxinidir}/tox.ini --cov={toxinidir}/skidl
+    py.test --ignore {toxinidir}/tests/examples -m "spice" --cov-config={toxinidir}/tox.ini --cov={toxinidir}/skidl
 
 [testenv:{clean,build}]
 description =
